refactor: improve period map. move utils into separate file with chart options
